Moved consensus package files into single consensus.py module
